Implement both enum-based and nan-boxed JsValue inner behind a flag
