+++ cli/Cargo.toml
@@ -28,6 +28,7 @@ dhat = { workspace = true, optional = true }
 [features]
 default = ["boa_engine/annex-b", "boa_engine/experimental", "boa_engine/intl_bundled"]
 dhat = ["dep:dhat"]
+nan-box-jsvalue = ["boa_engine/nan-box-jsvalue"]
 
 [target.x86_64-unknown-linux-gnu.dependencies]
 jemallocator.workspace = true
